software/requirements.txt
mss==10.1.0
requests==2.32.5
pystray==0.19.5
Pillow==11.3.0
pyinstaller==6.16.0